Uncommented out searchView. Had it off for testing since previews are all broken.
